+++ packages/hull-connector-framework/src/test-scenario.js
@@ -1,10 +1,14 @@
 // @flow
 import type { TestScenarioDefinition } from "./test-scenario-runner";
 
+const express = require("express");
 const TestScenarioRunner = require("./test-scenario-runner");
 
-function testScenario(scenarioDefinition: TestScenarioDefinition): Promise<*> {
-  const runner = new TestScenarioRunner(scenarioDefinition);
+function testScenario(
+  connectorServer: express => express,
+  scenarioDefinition: TestScenarioDefinition
+): Promise<*> {
+  const runner = new TestScenarioRunner(connectorServer, scenarioDefinition);
   return runner.run();
 }