+++ packages/hull/src/utils/promise-to-writable-stream.js
@@ -1,6 +1,6 @@
 // @flow
 const { Writable } = require("stream");
-const debug = require("debug")("promise-to-readable-stream");
+const debug = require("debug")("hull:promise-to-writable-stream");
 
 /**
  * A helper function which allows to easily pipe any stream into a promise.
@@ -15,9 +15,20 @@ function promiseToWritableStream(
     objectMode: true,
     write(chunk, encoding, callback) {
       debug("writing to a promise");
-      promise(chunk, encoding)
-        .then(() => callback())
-        .catch(error => callback(error));
+      try {
+        promise(chunk, encoding)
+          .then(() => {
+            debug("promise resolved");
+            callback();
+          })
+          .catch(error => {
+            debug("promise rejected", error);
+            callback(error);
+          });
+      } catch (error) {
+        debug("error thrown", error);
+        callback(error);
+      }
     }
   });
 }