+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v0.0.2-monorepo
+- [hull-typeform] more integration tests, verify ident from hidden fields.
+  updated documentation
+- [hull-node] brought back helpers, removed stuff from utils
+
 ## v0.0.1-monorepo
 - initial version of the monorepo
 - [hull-typeform] new version working with upgraded Typeform API